Update vtgate client
* Refactor StaticAuth options into a package.
* Update vtgate client to pass static auth options when present.
