+++ py/util/__init__.py
@@ -11,13 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""This is the setup script for the submodules in the Vitess python client.
-"""
-
-from distutils.core import setup
-
-
-setup(name="vitess",
-      packages=["vtctl", "vtdb", "vtproto", "vttest", "util"],
-      platforms="Any",
-     )
