+++ py/util/static_auth_client.py
@@ -1,23 +1,27 @@
 # Copyright 2017 Google Inc.
-# 
+#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-# 
+#
 #     http://www.apache.org/licenses/LICENSE-2.0
-# 
+#
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""This is the setup script for the submodules in the Vitess python client.
-"""
 
-from distutils.core import setup
+import json
 
+class StaticAuthClientCreds():
+    """Metadata wrapper for StaticAuthClientCreds."""
 
-setup(name="vitess",
-      packages=["vtctl", "vtdb", "vtproto", "vttest", "util"],
-      platforms="Any",
-     )
+    def __init__(self, auth_static_client_creds):
+        self._credentials = auth_static_client_creds
+        with open(self._credentials) as data_file:
+          self._data = json.load(data_file)
+
+    def metadata(self):
+        return (('username', self._data['Username']), ('password', self._data['Password']),)
+
